+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/items/explosivebarrel/ExplosiveBarrel.java
@@ -12,7 +12,7 @@ import net.minecraft.world.level.block.state.properties.NoteBlockInstrument;
 import net.minecraft.world.level.material.MapColor;
 
 @Label(name = "Explosive Barrel", description = "Add a new explosive block.")
-@LoadFeature(module = Modules.Ids.ITEMS)
+@LoadFeature(module = Modules.Ids.ITEMS, canBeDisabled = false)
 public class ExplosiveBarrel extends Feature {
 	public static final SimpleBlockWithItem BLOCK = SimpleBlockWithItem.register("explosive_barrel", () -> new ExplosiveBarrelBlock(BlockBehaviour.Properties.of().mapColor(MapColor.WOOD).ignitedByLava().instrument(NoteBlockInstrument.BASS).strength(2.5F).sound(SoundType.WOOD)));