+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/mining/quaron/Quaron.java
@@ -36,7 +36,7 @@ import net.minecraftforge.registries.RegistryObject;
 import java.util.EnumMap;
 
 @Label(name = "Quaron", description = "Add Quaron, a new metal made by alloying Iron and Amethyst")
-@LoadFeature(module = Modules.Ids.MINING)
+@LoadFeature(module = Modules.Ids.MINING, canBeDisabled = false)
 public class Quaron extends Feature {
 	public static final TagKey<Item> TOOL_EQUIPMENT = TagKey.create(Registries.ITEM, new ResourceLocation(IguanaTweaksExpanded.MOD_ID, "equipment/hand/tools/quaron"));